+++ node/src/Commands.ts
@@ -1279,6 +1279,13 @@ export function createExpireAt(
 /**
  * @internal
  */
+export function createExpireTime(key: string): command_request.Command {
+    return createCommand(RequestType.ExpireTime, [key]);
+}
+
+/**
+ * @internal
+ */
 export function createPExpire(
     key: string,
     milliseconds: number,
@@ -1304,6 +1311,13 @@ export function createPExpireAt(
             ? [key, unixMilliseconds.toString()]
             : [key, unixMilliseconds.toString(), option];
     return createCommand(RequestType.PExpireAt, args);
+}
+
+/**
+ * @internal
+ */
+export function createPExpireTime(key: string): command_request.Command {
+    return createCommand(RequestType.PExpireTime, [key]);
 }
 
 /**
@@ -1944,6 +1958,9 @@ function addTrimOptions(options: StreamTrimOptions, args: string[]) {
     }
 }
 
+/**
+ * @internal
+ */
 export function createXAdd(
     key: string,
     values: [string, string][],
@@ -1971,6 +1988,16 @@ export function createXAdd(
     });
 
     return createCommand(RequestType.XAdd, args);
+}
+
+/**
+ * @internal
+ */
+export function createXDel(
+    key: string,
+    ids: string[],
+): command_request.Command {
+    return createCommand(RequestType.XDel, [key, ...ids]);
 }
 
 /**
@@ -3011,6 +3038,16 @@ export function createRandomKey(): command_request.Command {
     return createCommand(RequestType.RandomKey, []);
 }
 
+/** @internal */
+export function createWatch(keys: string[]): command_request.Command {
+    return createCommand(RequestType.Watch, keys);
+}
+
+/** @internal */
+export function createUnWatch(): command_request.Command {
+    return createCommand(RequestType.UnWatch, []);
+}
+
 /**
  * This base class represents the common set of optional arguments for the SCAN family of commands.
  * Concrete implementations of this class are tied to specific SCAN commands (SCAN, HSCAN, SSCAN,
@@ -3054,4 +3091,98 @@ export function createZScan(
     }
 
     return createCommand(RequestType.ZScan, args);
+}
+
+/** @internal */
+export function createSetRange(
+    key: string,
+    offset: number,
+    value: string,
+): command_request.Command {
+    return createCommand(RequestType.SetRange, [key, offset.toString(), value]);
+}
+
+/**
+ * @internal
+ */
+export function createLMPop(
+    keys: string[],
+    direction: ListDirection,
+    count?: number,
+): command_request.Command {
+    const args: string[] = [keys.length.toString(), ...keys, direction];
+
+    if (count !== undefined) {
+        args.push("COUNT");
+        args.push(count.toString());
+    }
+
+    return createCommand(RequestType.LMPop, args);
+}
+
+/**
+ * @internal
+ */
+export function createBLMPop(
+    timeout: number,
+    keys: string[],
+    direction: ListDirection,
+    count?: number,
+): command_request.Command {
+    const args: string[] = [
+        timeout.toString(),
+        keys.length.toString(),
+        ...keys,
+        direction,
+    ];
+
+    if (count !== undefined) {
+        args.push("COUNT");
+        args.push(count.toString());
+    }
+
+    return createCommand(RequestType.BLMPop, args);
+}
+
+/**
+ * @internal
+ */
+export function createPubSubChannels(
+    pattern?: string,
+): command_request.Command {
+    return createCommand(RequestType.PubSubChannels, pattern ? [pattern] : []);
+}
+
+/**
+ * @internal
+ */
+export function createPubSubNumPat(): command_request.Command {
+    return createCommand(RequestType.PubSubNumPat, []);
+}
+
+/**
+ * @internal
+ */
+export function createPubSubNumSub(
+    channels?: string[],
+): command_request.Command {
+    return createCommand(RequestType.PubSubNumSub, channels ? channels : []);
+}
+
+/**
+ * @internal
+ */
+export function createPubsubShardChannels(
+    pattern?: string,
+): command_request.Command {
+    return createCommand(RequestType.PubSubSChannels, pattern ? [pattern] : []);
+}
+
+/**
+ * @internal
+ */
+export function createPubSubShardNumSub(
+    channels?: string[],
+): command_request.Command {
+    return createCommand(RequestType.PubSubSNumSub, channels ? channels : []);
 }