+++ node/tests/TestUtilities.ts
@@ -471,6 +471,7 @@ export async function transactionTest(
     const key21 = "{key}" + uuidv4(); // list for sort
     const key22 = "{key}" + uuidv4(); // list for sort
     const key23 = "{key}" + uuidv4(); // zset random
+    const key24 = "{key}" + uuidv4(); // list value
     const field = uuidv4();
     const value = uuidv4();
     // array of tuples - first element is test name/description, second - expected return value
@@ -478,6 +479,12 @@ export async function transactionTest(
 
     baseTransaction.publish("test_message", key1);
     responseData.push(['publish("test_message", key1)', 0]);
+    baseTransaction.pubsubChannels();
+    responseData.push(["pubsubChannels()", []]);
+    baseTransaction.pubsubNumPat();
+    responseData.push(["pubsubNumPat()", 0]);
+    baseTransaction.pubsubNumSub();
+    responseData.push(["pubsubNumSub()", {}]);
 
     baseTransaction.flushall();
     responseData.push(["flushall()", "OK"]);
@@ -505,6 +512,15 @@ export async function transactionTest(
     responseData.push(["echo(value)", value]);
     baseTransaction.persist(key1);
     responseData.push(["persist(key1)", false]);
+
+    if (gte(version, "7.0.0")) {
+        baseTransaction.expireTime(key1);
+        responseData.push(["expiretime(key1)", -1]);
+
+        baseTransaction.pexpireTime(key1);
+        responseData.push(["pexpiretime(key1)", -1]);
+    }
+
     baseTransaction.set(key2, "baz", { returnOldValue: true });
     responseData.push(['set(key2, "baz", { returnOldValue: true })', null]);
     baseTransaction.customCommand(["MGET", key1, key2]);
@@ -517,6 +533,8 @@ export async function transactionTest(
     responseData.push(["mget([key1, key2])", ["bar", "baz"]]);
     baseTransaction.strlen(key1);
     responseData.push(["strlen(key1)", 3]);
+    baseTransaction.setrange(key1, 0, "GLIDE");
+    responseData.push(["setrange(key1, 0, 'GLIDE'", 5]);
     baseTransaction.del([key1]);
     responseData.push(["del([key1])", 1]);
     baseTransaction.hset(key4, { [field]: value });
@@ -546,6 +564,24 @@ export async function transactionTest(
         field + "4",
     ]);
     responseData.push(["lpush(key5, [1, 2, 3, 4])", 4]);
+
+    if (gte("7.0.0", version)) {
+        baseTransaction.lpush(key24, [field + "1", field + "2"]);
+        responseData.push(["lpush(key22, [1, 2])", 2]);
+        baseTransaction.lmpop([key24], ListDirection.LEFT);
+        responseData.push([
+            "lmpop([key22], ListDirection.LEFT)",
+            { [key24]: [field + "2"] },
+        ]);
+        baseTransaction.lpush(key24, [field + "2"]);
+        responseData.push(["lpush(key22, [2])", 2]);
+        baseTransaction.blmpop([key24], ListDirection.LEFT, 0.1, 1);
+        responseData.push([
+            "blmpop([key22], ListDirection.LEFT, 0.1, 1)",
+            { [key24]: [field + "2"] },
+        ]);
+    }
+
     baseTransaction.lpop(key5);
     responseData.push(["lpop(key5)", field + "4"]);
     baseTransaction.llen(key5);
@@ -845,6 +881,8 @@ export async function transactionTest(
         'xtrim(key9, { method: "minid", threshold: "0-2", exact: true }',
         1,
     ]);
+    baseTransaction.xdel(key9, ["0-3", "0-5"]);
+    responseData.push(["xdel(key9, [['0-3', '0-5']])", 1]);
     baseTransaction.rename(key9, key10);
     responseData.push(["rename(key9, key10)", "OK"]);
     baseTransaction.exists([key10]);